Support fixed-width integer types
This adds support for signed and unsigned types of fixed size:

    int8, int16, int32, int64, as well as
    uint8, uint16, uint32, uint64

mapping them to exsting types from small to hyper. The keywords are
specified by OMG (IDL 4.2) and implemented by both MS and Samba.

Note that this conjures a conflict with the existing builtin attributes
`int32' and `int64' (introduced with commit ad8af5bcb2). As a
workaround, the parser has been adapted to treat these two keywords as
identifiers in attribute context.